+++ include/create.hpp
@@ -3,4 +3,3 @@
 #include <string>
 
 void createBox(std::string id);
-void waitForInput();
