+++ source/common/util.cpp
@@ -1,0 +1,8 @@
+#include <3ds.h>
+#include "common/util.hpp"
+
+void waitForInput()
+{
+    while (aptMainLoop() && !hidKeysDown()) hidScanInput();
+    hidScanInput();
+}
